Repeat pthread documentation, instead of linking to C functions.
First, the C function links are redundant, since it is obvious which
function the binding pertains to from the name. Second, pthread is
indeed a low-level binding, but that doesn't mean the fact that binds
to the C interface under the hood should leak into the documentation.
Third, even other low-level binding libraries, like Unix, don't
normally link to the corresponding C function.